CRITICAL BUG FIX There was a buggy design (!2) in the way property wrapper classes were keeping track of the property name and the model they belong to.
As a fact (1) as properties are class-level members (vs
instance-level), they got shared among the nheritance chain, from base
down to derivated classes. So in essence if a property got wrapped
(automatically or manually, as in the case of signals), the wrapping
class instance has to be associated to all the models in the hierarchy
chain, and not to the last one registering it.

Now instead of associating to a property name and a model, each
wrapping instance is associated to a set of pairs (model,
property-name).

WARNING: This is still a partial commitment. Further analysis has to
be carried out, to understand if the solution captures all the
desidered behaviour, and for checking non-regression.

Thanks to Tobias Weber for pinpointing this issue.

(1) Even if the bug was afflicting all kind of property wrappers (like
signals, lists, maps and user classes), this problem was occurring
specifically only for signals, but only due to a technical reason.

(!2) This was a trivial design issue that could be avoided with better
design analysis.

[RC]