FEATURE
Added partial support for logical properties (not completed, the
commit is partial, but it is not expected to break anything).

At the moment, I added only support for two new decorator: 'getter'
and 'setter' which can be used to decorate methods in Models according
to what explained in ticket:7

Added also an initial set of automatic tests for logical properties.
The test at the moment checks that the new decorators do their work
correctly.

Next step: complete support for logical props at metaclass level,
exploiting annotations made by the new decorators.

IMPORTANT TO BE DONE: 
1. Check all chosen names ('getter', 'setter') and
2. Check the place they are curreblty located (model.py) is the right
place.
3. Documentation to be updated:
3.1 rename custom props to Logical props.
3.2 document the syntax and semantics of the new decorators