BUG FIX Fixed a bug regarding ticket:29.
Now (prop_name, method) are associated to the kwargs.
This required to change the interfce of new methods in Observer.

I also added a constraint (already documented): it is not possible to declare that a method is a notification for property 'prop' more than one times.

For example
 @Observer.observe("prop", ...)
 @Observer.observe("prop", ...)
 def foo(self, ...)

...Raises ValueError exception.

Now all tests in observer.py pass.