FEATURE Implemented (still partial) dependencies OPs. (Discussed in ticket:9)
The feature is still incomplete: in particular checks that there are no 
circular dependencies have to be done.
The feature should be in place, but testing has to be done. 

This commit is expected to do not break existing code.
Notice that an existing test had to be modified, as it was not correctly 
working otherwise after the changes. 

Updated version to 1.99.2