FEATURE
Added the possibility to raise ValueError in adapter widget getter
function.  When the widget value getter is called to retrieve the
value hold by the widget, it is now possible to have the getter raise
ValueError when the value hold in the widget for any reason cannot be
retrieved, or if in general it is not valid.

When the adapter detects that ValueError has been raised, it cancel
the operation, with the effect of not updating the adapted OP (which
will keep its current value).

This is useful to enable adaptation of widgets whose value is valid
(or should be considered) or not depending on the widget's state.