FEATURE
Added support for adapting RadioButton and RadioAction.

It is now possible to adapt a (group of) radio buttons or actions to
one OP. The value of the OP is a string containing the label of the
button/action which is currently active. 

Adaptation is bi-directional, meaning that when the OP is set to a
value, if that value matches the label of one of the adapted radio
buttons/actions, than that button/action will be activated. If no
match is found, the radio buttons/actions are kept untouched.

See the attached test for an example (the test has to be automatized).

This feature has to be documented in the manual.