TEST the View design proposal from #3.
This is a fixed niew_07.py and a port of mini-yoman from 1.99.1 (since the
current version is broken up into multiple GtkBuilder files)

The UI was only loaded into Glade and saved out as GtkBuilder, not other
changes.

The handler argument to all controllers wil go away once we adjust Controller
to no longer expect a glade_xmlWidgets attribute.